# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Mealswapp API
  version: 0.3.0
  description: Shared gateway source of truth for Phase 03 authentication and gateway contracts.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/v1/auth/register:
    post:
      operationId: postApiV1AuthRegister
      summary: Register with email, password, and current consent versions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/AuthSession"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/auth/login:
    post:
      operationId: postApiV1AuthLogin
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/AuthSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      operationId: postApiV1AuthLogout
      summary: Revoke the current refresh session and clear auth cookies
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Session cleared.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/auth/refresh:
    post:
      operationId: postApiV1AuthRefresh
      summary: Rotate the refresh token and issue a fresh access token
      security:
        - refreshCookie: []
      responses:
        "200":
          $ref: "#/components/responses/AuthSession"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/verify-email:
    post:
      operationId: postApiV1AuthVerifyEmail
      summary: Mark the authenticated email login method verified
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "200":
          description: Verified-login projection update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/auth/password-reset/request:
    post:
      operationId: postApiV1AuthPasswordResetRequest
      summary: Request a password reset without revealing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Generic password-reset request acceptance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequestEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/auth/password-reset/consume:
    post:
      operationId: postApiV1AuthPasswordResetConsume
      summary: Consume a single-use reset token and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConsumeRequest"
      responses:
        "200":
          description: Password reset completion.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetConsumeEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mealswapp_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: mealswapp_refresh
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
...
    AuthSession:
      description: Authenticated session envelope. Access and refresh token values are set only as HttpOnly cookies.
      headers:
        Set-Cookie:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthSessionEnvelope"
    Error:
      description: User-safe gateway error envelope with no client-visible audit snapshot, raw provider payload, or internal diagnostics.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    # Implements DESIGN-004 JobStatusTracker and DESIGN-010 RateLimiter shared admission contract.
...
  schemas:
    AppError:
      type: object
      additionalProperties: false
      required: [category, code, message, retryable]
      properties:
        category:
          type: string
          enum: [validation, auth, entitlement, security, network, timeout, server, dependency, rate_limit, unknown]
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 1
          maxLength: 120
    # Implements DESIGN-009 AdminController audit-safe error boundary.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, error]
      properties:
        status:
          type: string
          const: error
        requestId:
          type: string
          maxLength: 120
        error:
          $ref: "#/components/schemas/AppError"
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, privacyPolicyVersion, termsVersion]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        privacyPolicyVersion:
          type: string
        termsVersion:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetConsumeRequest:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 32
        newPassword:
          type: string
          minLength: 12
    AuthSessionData:
      type: object
      required: [userId, role, hasVerifiedLoginMethod, accessExpiresAt, refreshExpiresAt]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, admin]
        hasVerifiedLoginMethod:
          type: boolean
        accessExpiresAt:
          type: string
          format: date-time
        refreshExpiresAt:
          type: string
          format: date-time
    AuthSessionEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          $ref: "#/components/schemas/AuthSessionData"
...
    VerifyEmailEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          required: [hasVerifiedLoginMethod]
          properties:
            hasVerifiedLoginMethod:
              type: boolean
              const: true
    PasswordResetRequestEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          required: [accepted]
          additionalProperties: false
          properties:
            accepted:
              type: boolean
              const: true
    PasswordResetConsumeEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          required: [reset]
          properties:
            reset:
              type: boolean
              const: true
...